app/gts-service/requirements.txt
fastapi==0.83.0
uvicorn==0.18.3
databases==0.6.1
ormar==0.11.3
asyncpg==0.26.0
psycopg2-binary==2.9.3
httpx==0.23.0
aiohttp==3.8.1
python-socketio==5.7.2
